==> modeling_notebooks/00_install_requirements.ipynb <==
!virtualenv -p python3 ../venv_fantasy

!source ../venv_fantasy/bin/activate

!pip3 install -r ../requirements.txt

# Download corpora for sentiment analysis
!python3 -m textblob.download_corpora

!ipython kernel install --user --name=venv_fantasy
