==> modeling_notebooks/01_pull_data.ipynb <==
# Compile Raw Project Data

This project uses data from four sources:

1. Fantasy statisics for Tom Brady and LeSean McCoy from the 2010 - 2012 seasons. Downloaded by hand from profootballreference.com ("Tom Brady NFL Fantasy Football Statistics". Pro-Football-Reference.com. Retrieved October 19, 2019; "LeSean McCoy NFL Fantasy Football Statistics". Pro-Football-Reference.com. Retrieved October 19, 2019.) Located in ~/data/data_raw/players.

2. Game data for all NFL games played since 1966. Used for weather and betting features. Downloaded from Kaggle (spreadspoke (2019, August). NFL Scores and Betting Data. Retrieved October 19, 2019.). Located in ~/data/data_raw/games. 

3. Defensive numbers put up by every team for all games in the 2010 - 2012 seasons. Downloaded using the Python sportsreference API. Located in ~/data/data_raw/teams. 

4. NFL tweets from the 2010 - 2012 seasons. Downloaded from Sinha, et. al. ( Predicting the NFL using Twitter. Shiladitya Sinha, Chris Dyer, Kevin Gimpel, and Noah A. Smith. In Proceedings of the ECML/PKDD 2013 Workshop on Machine Learning and Data Mining for Sports Analytics, Prague, Czech Republic, September 2013.) Located in ~/data/data_raw/tweets_meta. 

    The former two datasets were manually downloaded and are available by default in the repository. The defensive data is also present in the repository, but was downloaded using the script located at ~/data_compilation/get_defensive_stats.py. Finally, the Twitter dataset is available in the repository, but it does not include any of the tweet text, per Twitter's terms of service. The script ~/data_compilation/pull_tweet_text.py will use the tweepy package to retrieve the text for each tweet and write it to the ~/data/data_raw/tweets_text folder. 
    
All of this data will be required for the following notebooks. This notebook will execute the two scripts.


#### Defensive Data

**UPDATE**: As of 11/9/2019, there seems to be an error with the NFL section of the sportsreference Python API. I am therefore commenting out the next line that attempts to pull data using the API. Following notebooks should still work because the data pulled by the script is included with the repository.

# # Defensive Stats
# !python3 ../data_compilation/get_defensive_stats.py

#### Twitter Text


Credit for this script: https://stackoverflow.com/questions/28384588/twitter-api-get-tweets-with-specific-id


# Twitter Data
!python3 ../data_compilation/pull_tweet_text.py
