==> modeling_notebooks/02_clean_game_data.ipynb <==
# Clean Player Data and Join Game Info

This is the first of three data cleaning notebooks. We will read in each raw player .csv and perform the following steps:

1. Drop unnecessary columns
2. Rename confusing columns
3. Transform the 'home/away' column to a Boolean indiciating if the game was played at home.

We will then read in game data and join relevant information about games our players played in to the base player DataFrames. We'll perform the following cleaning steps on game data before joining:
1. Filter out games not involving one of our players' teams
2. Drop unneeded columns
3. Transform 'team_favorite_id' column into a boolean indicating whether or not the player's team was favored. 

Finally, we will merge the DataFrames on game date and team.

### Setup

import os
import pandas as pd

### Players

# Read in data
players = {}
path = '../data/data_raw/players'
for filename in os.listdir(path):
    if '.csv' in filename:
        players[filename.split('.')[0]] = pd.read_csv(os.path.join(path,filename))

for key, value in players.items():
    print(value.head())

# Define data types; drop columns we don't need; rename columns for ease of uses
unneeded_cols = [
    'Tm', 
    'Result', 
    'Pos'
]
renamed_cols = {'G#':'week','home/away':'home'} # Convert this to boolean

for key, value in players.items():
    value.drop(unneeded_cols, axis=1, inplace=True)
    value.rename(columns=renamed_cols, inplace=True)
    value.columns = [col.lower() for col in value.columns]
    value['date'] = pd.to_datetime(value['date'], errors='coerce')
    value['home'] = [1 if pd.isnull(loc) else 0 for loc in value['home']]
    
for key, value in players.items():
    print(value.head())

### Games

games = pd.read_csv('../data/data_raw/games/game_metadata.csv')
games.head()

We can't directly join these dataframes because 'games' uses full team name while the 'players' dfs use abbreviation. Since we only have two teams, we can just filter on team name and join by date. To scale this in the future, we would want to build out this process so we can directly join.

games['schedule_date'] = pd.to_datetime(games['schedule_date'], errors='coerce')

games_patriots = games.loc[(games['team_home']=='New England Patriots') | (games['team_away']=='New England Patriots')]
games_eagles = games.loc[(games['team_home']=='Philadelphia Eagles') | (games['team_away']=='Philadelphia Eagles')]

print(games_patriots.shape)
print(games_eagles.shape)

for key, value in players.items():
    if key =='brady':
        players[key] = value.merge(games_patriots, how='inner', left_on='date', right_on='schedule_date')
    else:
        players[key] = value.merge(games_eagles, how='inner', left_on='date',  right_on='schedule_date')

# More column manipulation
unneeded_cols = [
    'schedule_date', 
    'schedule_season', 
    'schedule_week', 
    'schedule_playoff',
    'team_home',
    'score_home',
    'team_away',
    'score_away',
    'stadium',
    'stadium_neutral'
]
renamed_cols = {'team_favorite_id':'favorite'}

for key, value in players.items():
    value.drop(unneeded_cols, axis=1, inplace=True)
    value.rename(columns=renamed_cols, inplace=True)
    if key=='brady':
        team_id = 'NE'
    else:
        team_id = 'PHI'
    value['favorite'] = [1 if code==team_id else 0 for code in value['favorite']]
    
for key, value in players.items():
    print(value.head())

# Save data to new folder so we can use it easily
for key, value in players.items():
    value.to_csv('../data/data_modified/players/{}.csv'.format(key), index=False)
