==> modeling_notebooks/03_clean_defensive_data.ipynb <==
# Clean Opponent Defensive Data

The primary task in this notebook is to ensure consistency of abbreviations between defense and players dataframes. We will find inconsistencies between the abbreviations used by our defensive data and our base player data and map correct them so that we can join later.

import os
import pandas as pd

# Read in  player data
players = {}
path = '../data/data_modified/players'
for file in os.listdir(path):
    if file.split('.')[1]=='csv':
        players[file.split('.')[0]] = pd.read_csv(os.path.join(path,file))


# Read in defense data
defense = pd.read_csv('../data/data_raw/teams/defensive_stats.csv')
defense.head()

# Check to make sure abbreviation conventions are the same
players_abbrs = []
for key, value in players.items():
    players_abbrs.extend(value['opp'].tolist())

sum(~pd.Series(players_abbrs).isin(defense['team'].unique()))

# Figure out which cases are different
print(pd.Series(players_abbrs).loc[~pd.Series(players_abbrs).isin(defense['team'].unique())].unique())
print(pd.Series(defense['team'].unique()).loc[~pd.Series(defense['team'].unique()).isin(players_abbrs)].unique())

# Create mappings to convert inconsistent abbreviations. This is just based on domain knowledge.
# Again, to scalably deploy this on multiple players we would need to automate this process.
abbr_mappings = {
    'OTI':'TEN',
    'CLT':'IND',
    'HTX':'HOU',
    'RAM':'STL',
    'CRD':'ARI',
    'RAV':'BAL',
    'RAI':'OAK'}
defense.replace(abbr_mappings, inplace=True)
sum(~pd.Series(players_abbrs).isin(defense['team'].unique()))

defense.to_csv('../data/data_modified/teams/defensive_stats.csv', index=False)
