==> modeling_notebooks/04_clean_twitter_data.ipynb <==
# Clean Twitter Data

We have two tasks in this notebook:
1. Ensure consistency of abbreviations between Twitter data and players data
2. Join twitter text and meta dataframes and label columns


import os
import pandas as pd

# Read in text data
tweet_text = pd.DataFrame()
path = '../data/data_raw/tweets_text' # Note that the user needs to run 01_pull_data.ipynb for this to work
for filename in os.listdir(path):
    if '.csv' in filename:
        tweet_text = tweet_text.append(pd.read_csv(os.path.join(path,filename), header=None))
tweet_text.columns=['tweet_id', 'content']
tweet_text.head()

# Read in and label columns for metadata
tweet_meta = pd.DataFrame()
path = '../data/data_raw/tweets_meta'
col_key_file = 'tweets.key.csv'
colnames = pd.read_csv(os.path.join(path,col_key_file)).columns.tolist()
for filename in os.listdir(path):
    if ('.csv' in filename) & (filename != col_key_file):
        year_search = re.search(r'\d\d\d\d', filename)
        tweet_meta = tweet_meta.append(pd.read_csv(os.path.join(path,filename), header=None, names=colnames))

tweet_meta.head()

# Merge the data
tweets_all = tweet_meta.merge(tweet_text, how='left', on='tweet_id')
print('{} tweets deleted since 2012. {} tweets remaining.'.format(len(tweet_meta) - len(tweet_text), len(tweet_text)))
print(tweets_all.head())

# Read in defensive stats for team abrreviation comparison
defensive_stats = pd.read_csv('../data/data_modified/teams/defensive_stats.csv')
print(defensive_stats['team'][~defensive_stats['team'].isin(tweets_all['team'])].unique())
print(tweets_all['team'][~tweets_all['team'].isin(defensive_stats['team'])].unique())

# Manually map names using domain knowledge
abbr_mappings = {
    'NE':'NWE',
    'GB':'GNB',
    'SF':'SFO',
    'SD':'SDG',
    'TB':'TAM',
    'NO':'NOR',
    'KC':'KAN'}
tweets_all.replace(abbr_mappings, inplace=True)
print(sum(~tweets_all['team'].isin(defensive_stats['team'].unique())))
print(sum(~tweets_all['opponent'].isin(defensive_stats['team'].unique())))

# Remove escape chars
tweets_all['content'] = [tweet.replace('\r', ' ').replace('\n', ' ') for tweet in tweets_all['content']]

tweets_all.to_csv('../data/data_modified/tweets/tweets.csv', index=False)
