==> modeling_notebooks/05_sentiment_analysis.ipynb <==
# Sentiment Analysis

In this notebook, we will use the tweet-preprocessor package to clean our tweet text. We'll drop non-english tweets, and then use Stanford Core NLP for sentiment analysis.

from langdetect import detect
import pandas as pd
import preprocessor as p
from string import punctuation
import time
from pycorenlp import StanfordCoreNLP

#### Read in data

tweets = pd.read_csv('../data/data_modified/tweets/tweets.csv')

tweets['content'][0:10].tolist()

#### Clean data using the preprocessor package.

%%time
# Clean all the tweets, removing leading and lagging punctuation
tweets['content_clean'] = [p.clean(tweet).strip(punctuation) for tweet in tweets['content'].tolist()]

#### Find non-English tweets

%%time
# Find how many tweets are not in English using langdetect
langs = []
for text in tweets['content_clean']:
    try:
        langs.append(detect(text))
    except:
        langs.append('')
tweets['language'] = langs

# That was a long process, so let's save our progress
tweets.to_csv('../data/data_modified/tweets/tweets.csv', index=False)

Checking the data anecdotally, it looks like most non-English tweets that are not Spanish are either false positives or empty (i.e, only hashtags). We want to keep all tweets for count features. We only want to do sentiment analysis for non-Spanish, non-empty tweets.

For the purposes of this model, we will treat all 'empty text' tweets as 'neutral' sentiment. This is contestable, however - anecdotal evidence suggests these may actually be positive (most are hashtag-only support for a given team). We won't dive into this much here, but it is a clear area for further research.

### Sentiment Analysis

# Activate Stanford Core NLP - must first download English version from https://stanfordnlp.github.io/CoreNLP/
# and run this command locally in the unzipped folder

# !java -mx5g -cp "*" edu.stanford.nlp.pipeline.StanfordCoreNLPServer -timeout 10000

%%time
# Use NLP on the sentence level to determine overall sentiment of the tweet
try:
    tweets = pd.read_csv('../data/data_modified/tweets/tweets_sentiment.csv') # if the sentiment file already exists, read it in.
except:
    nlp = StanfordCoreNLP('http://localhost:9000')

    sentiment_mappings = {
        'Verypositive': 2,
        'Positive': 1,
        'Negative': -1,
        'Verynegative': -2,
        'Neutral': 0
    }
    sentiments = []

    for i, tweet in enumerate(tweets['content_clean']):
        parsed_tweet = nlp.annotate(tweet,
                           properties={
                               'annotators': 'sentiment',
                               'outputFormat': 'json',
                               'timeout': 100000,
                           })
        sentiment = 0
        for s in parsed_tweet["sentences"]:
            sentiment += sentiment_mappings[s["sentiment"]]
        sentiments.append(sentiment)
    tweets['sentiment'] = sentiments
    tweets['sentiment'] = [0 if sent[0]=='es' else sent[1] for sent in zip(tweets['language'], tweets['sentiment'])]
    tweets.loc[pd.isnull(tweets['sentiment']),['sentiment']] = 0 # Assume sentiment of 0 for deleted tweets
    tweets.to_csv('../data/data_modified/tweets/tweets_sentiment.csv', index=False)
