==> modeling_notebooks/14_statistical_comparison.ipynb <==
# Statistical Comparison of Model Results

The purpose of this notebook is to do a statistical comparison of our baseline and final LSTM models, and our moving average baseline and final LSTM model. We want to see if the improvement (or lack thereof) of the final model is statistically significant.

import pickle
from scipy.stats import ttest_rel, ttest_1samp

### Tom Brady

baseline_lstm = pickle.load(open('../data/prediction_outputs/brady_baseline_lstm_maes', 'rb'))
final_lstm = pickle.load(open('../data/prediction_outputs/brady_final_lstm_maes', 'rb'))

# Just hard code moving average predicition from notebook 9
baseline_moving_avg = 6.33

# Baseline LSTM vs Final LSTM
ttest_rel(baseline_lstm, final_lstm)

# Moving Average vs Final LSTM
ttest_1samp(final_lstm, baseline_moving_avg)

### LeSean McCoy

baseline_lstm = pickle.load(open('../data/prediction_outputs/mccoy_baseline_lstm_maes', 'rb'))
final_lstm = pickle.load(open('../data/prediction_outputs/mccoy_final_lstm_maes', 'rb'))

# Just hard code moving average predicition from notebook 9
baseline_moving_avg = 2.65

# Baseline LSTM vs Final LSTM
ttest_rel(baseline_lstm, final_lstm)

# Moving Average vs Final LSTM
ttest_1samp(final_lstm, baseline_moving_avg)
